Re-init git repo to clean up the data noise.
Had added the USGS geonames dataset to the R pkg but then
realized it takes far too long to re-install the pkg during
development with that data present. Rolling back the commits
was going to take more effort than this very early version
of the pkg warrants, so starting over.